projects/modules/10-django-fullstack/requirements.txt
django>=5.0
djangorestframework>=3.14
